move api.Client to client_test.go
