+++ node/api/client_test.go
@@ -1,54 +1,114 @@
 package api
 
 import (
-	"testing"
+	"encoding/json"
+	"errors"
+	"io"
+	"io/ioutil"
+	"net/http"
+	"strings"
 )
 
-// TestApiClient tests that the API client connects to the server tester and
-// can call and decode routes correctly.
-func TestApiClient(t *testing.T) {
-	if testing.Short() {
-		t.SkipNow()
-	}
-	t.Parallel()
-	st, err := createServerTester(t.Name())
-	if err != nil {
-		t.Fatal(err)
-	}
-	defer st.server.panicClose()
+// Client holds fields to make requests to a Sia API.
+type Client struct {
+	address  string
+	password string
+}
 
-	c := NewClient(st.server.listener.Addr().String(), "")
-	var gatewayInfo GatewayGET
-	err = c.Get("/gateway", &gatewayInfo)
-	if err != nil {
-		t.Fatal(err)
+// NewClient creates a new api.Client using the provided address and password.
+// If password is not the empty string, HTTP basic authentication will be used
+// to communicate with the API.
+func NewClient(address string, password string) *Client {
+	return &Client{
+		address:  address,
+		password: password,
 	}
 }
 
-// TestAuthenticatedApiClient tests that the API client connects to an
-// authenticated server tester and can call and decode routes correctly, using
-// the correct password.
-func TestAuthenticatedApiClient(t *testing.T) {
-	if testing.Short() {
-		t.SkipNow()
-	}
-	t.Parallel()
-	testpass := "testPassword"
-	st, err := createAuthenticatedServerTester(t.Name(), testpass)
+// Get requests the resource at `resource` and decodes it into `obj`, returning an
+// error if requesting or decoding the resource fails.  A non-2xx status code
+// constitutes a request failure.
+func (c *Client) Get(resource string, obj interface{}) error {
+	url := "http://" + c.address + resource
+	req, err := http.NewRequest("GET", url, nil)
 	if err != nil {
-		t.Fatal(err)
+		return err
 	}
-	defer st.server.panicClose()
+	req.Header.Set("User-Agent", "Sia-Agent")
+	if c.password != "" {
+		req.SetBasicAuth("", c.password)
+	}
+	res, err := http.DefaultClient.Do(req)
+	if err != nil {
+		return err
+	}
+	defer func() {
+		// res.Body should always be fully read even when discarding its content,
+		// such that the underlying connection can be reused.
+		io.Copy(ioutil.Discard, res.Body)
+		res.Body.Close()
+	}()
 
-	c := NewClient(st.server.listener.Addr().String(), "")
-	var walletAddress WalletAddressGET
-	err = c.Get("/wallet/address", &walletAddress)
-	if err == nil {
-		t.Fatal("api.Client did not return an error when requesting an authenticated resource without a password")
+	if res.StatusCode == http.StatusNotFound {
+		return errors.New("API call not recognized: " + resource)
 	}
-	c = NewClient(st.server.listener.Addr().String(), testpass)
-	err = c.Get("/wallet/address", &walletAddress)
+
+	// Decode the body as an Error and return this error if the status code is
+	// not 2xx.
+	if res.StatusCode < 200 || res.StatusCode > 299 {
+		var apiErr Error
+		err = json.NewDecoder(res.Body).Decode(&apiErr)
+		if err != nil {
+			return err
+		}
+		return apiErr
+	}
+
+	if res.StatusCode != http.StatusNoContent && obj != nil {
+		return json.NewDecoder(res.Body).Decode(obj)
+	}
+	return nil
+}
+
+// Post makes a POST request to the resource at `resource`, using `data` as the
+// request body.  The response, if provided, will be decoded into `obj`.
+func (c *Client) Post(resource string, data string, obj interface{}) error {
+	url := "http://" + c.address + resource
+	req, err := http.NewRequest("POST", url, strings.NewReader(data))
 	if err != nil {
-		t.Fatal(err)
+		return err
 	}
+	req.Header.Set("User-Agent", "Sia-Agent")
+	req.Header.Set("Content-Type", "application/x-www-form-urlencoded")
+	if c.password != "" {
+		req.SetBasicAuth("", c.password)
+	}
+	res, err := http.DefaultClient.Do(req)
+	if err != nil {
+		return err
+	}
+	defer func() {
+		// res.Body should always be fully read even when discarding its content,
+		// such that the underlying connection can be reused.
+		io.Copy(ioutil.Discard, res.Body)
+		res.Body.Close()
+	}()
+
+	if res.StatusCode == http.StatusNotFound {
+		return errors.New("API call not recognized: " + resource)
+	}
+
+	if res.StatusCode < 200 || res.StatusCode > 299 {
+		var apiErr Error
+		err = json.NewDecoder(res.Body).Decode(&apiErr)
+		if err != nil {
+			return err
+		}
+		return apiErr
+	}
+
+	if res.StatusCode != http.StatusNoContent && obj != nil {
+		return json.NewDecoder(res.Body).Decode(&obj)
+	}
+	return nil
 }
